:recycle: Move functions test and debug
