+++ src/Cool/DidYouKnow/testDidYouKnowModal.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include "Cool/DidYouKnow/didYouKnowModal.hpp"
+
+void test_did_you_know(DidYouKnowModal& _did_you_know);
